+++ proxy/frontend/type/mysql/src/main/java/org/apache/shardingsphere/proxy/frontend/mysql/command/query/text/query/MySQLMultiStatementsHandler.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.proxy.frontend.mysql.command.query.text.query;
 
+import org.apache.shardingsphere.driver.executor.engine.batch.preparedstatement.BatchExecutionUnit;
+import org.apache.shardingsphere.driver.executor.engine.batch.preparedstatement.BatchPreparedStatementExecutor;
 import org.apache.shardingsphere.infra.binder.context.statement.SQLStatementContext;
 import org.apache.shardingsphere.infra.binder.engine.SQLBindEngine;
 import org.apache.shardingsphere.infra.config.props.ConfigurationPropertyKey;
@@ -28,11 +30,8 @@ import org.apache.shardingsphere.infra.executor.kernel.model.ExecutionGroupConte
 import org.apache.shardingsphere.infra.executor.kernel.model.ExecutionGroupReportContext;
 import org.apache.shardingsphere.infra.executor.sql.context.ExecutionContext;
 import org.apache.shardingsphere.infra.executor.sql.context.ExecutionUnit;
-import org.apache.shardingsphere.infra.executor.sql.execute.engine.ConnectionMode;
-import org.apache.shardingsphere.infra.executor.sql.execute.engine.SQLExecutorExceptionHandler;
 import org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutionUnit;
 import org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutor;
-import org.apache.shardingsphere.infra.executor.sql.execute.engine.driver.jdbc.JDBCExecutorCallback;
 import org.apache.shardingsphere.infra.executor.sql.execute.result.update.UpdateResult;
 import org.apache.shardingsphere.infra.executor.sql.prepare.driver.DriverExecutionPrepareEngine;
 import org.apache.shardingsphere.infra.executor.sql.prepare.driver.jdbc.JDBCDriverType;
@@ -40,7 +39,6 @@ import org.apache.shardingsphere.infra.executor.sql.prepare.driver.jdbc.Statemen
 import org.apache.shardingsphere.infra.hint.HintValueContext;
 import org.apache.shardingsphere.infra.hint.SQLHintUtils;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
-import org.apache.shardingsphere.infra.metadata.database.resource.ResourceMetaData;
 import org.apache.shardingsphere.infra.metadata.database.rule.RuleMetaData;
 import org.apache.shardingsphere.infra.parser.SQLParserEngine;
 import org.apache.shardingsphere.infra.rule.ShardingSphereRule;
@@ -53,22 +51,21 @@ import org.apache.shardingsphere.proxy.backend.context.BackendExecutorContext;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
 import org.apache.shardingsphere.proxy.backend.handler.ProxyBackendHandler;
 import org.apache.shardingsphere.proxy.backend.response.header.ResponseHeader;
+import org.apache.shardingsphere.proxy.backend.response.header.update.MultiStatementsUpdateResponseHeader;
 import org.apache.shardingsphere.proxy.backend.response.header.update.UpdateResponseHeader;
 import org.apache.shardingsphere.proxy.backend.session.ConnectionSession;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.SQLStatement;
+import org.apache.shardingsphere.sql.parser.statement.core.statement.dml.InsertStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dml.UpdateStatement;
 
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.sql.Statement;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import java.util.regex.Pattern;
 
 /**
@@ -76,33 +73,41 @@ import java.util.regex.Pattern;
  */
 public final class MySQLMultiStatementsHandler implements ProxyBackendHandler {
     
+    private static final Pattern MULTI_INSERT_STATEMENTS = Pattern.compile(";(?=\\s*insert)", Pattern.CASE_INSENSITIVE);
+    
     private static final Pattern MULTI_UPDATE_STATEMENTS = Pattern.compile(";(?=\\s*update)", Pattern.CASE_INSENSITIVE);
     
     private static final Pattern MULTI_DELETE_STATEMENTS = Pattern.compile(";(?=\\s*delete)", Pattern.CASE_INSENSITIVE);
-    
-    private final KernelProcessor kernelProcessor = new KernelProcessor();
-    
-    private final JDBCExecutor jdbcExecutor;
-    
-    private final ConnectionSession connectionSession;
-    
-    private final SQLStatement sqlStatementSample;
     
     private final MetaDataContexts metaDataContexts = ProxyContext.getInstance().getContextManager().getMetaDataContexts();
     
     private final Collection<QueryContext> multiSQLQueryContexts = new LinkedList<>();
     
+    private final ConnectionSession connectionSession;
+    
+    private final SQLStatement sqlStatementSample;
+    
+    private final BatchPreparedStatementExecutor batchExecutor;
+    
     public MySQLMultiStatementsHandler(final ConnectionSession connectionSession, final SQLStatement sqlStatementSample, final String sql) {
-        jdbcExecutor = new JDBCExecutor(BackendExecutorContext.getInstance().getExecutorEngine(), connectionSession.getConnectionContext());
         connectionSession.getDatabaseConnectionManager().handleAutoCommit();
         this.connectionSession = connectionSession;
         this.sqlStatementSample = sqlStatementSample;
-        Pattern pattern = sqlStatementSample instanceof UpdateStatement ? MULTI_UPDATE_STATEMENTS : MULTI_DELETE_STATEMENTS;
+        JDBCExecutor jdbcExecutor = new JDBCExecutor(BackendExecutorContext.getInstance().getExecutorEngine(), connectionSession.getConnectionContext());
+        batchExecutor = new BatchPreparedStatementExecutor(metaDataContexts.getMetaData().getDatabase(connectionSession.getUsedDatabaseName()), jdbcExecutor, connectionSession.getProcessId());
+        Pattern pattern = getPattern(sqlStatementSample);
         SQLParserEngine sqlParserEngine = getSQLParserEngine();
         for (String each : extractMultiStatements(pattern, sql)) {
             SQLStatement eachSQLStatement = sqlParserEngine.parse(each, false);
             multiSQLQueryContexts.add(createQueryContext(each, eachSQLStatement));
         }
+    }
+    
+    private Pattern getPattern(final SQLStatement sqlStatementSample) {
+        if (sqlStatementSample instanceof InsertStatement) {
+            return MULTI_INSERT_STATEMENTS;
+        }
+        return sqlStatementSample instanceof UpdateStatement ? MULTI_UPDATE_STATEMENTS : MULTI_DELETE_STATEMENTS;
     }
     
     private SQLParserEngine getSQLParserEngine() {
@@ -134,34 +139,31 @@ public final class MySQLMultiStatementsHandler implements ProxyBackendHandler {
         return executeMultiStatements(prepareEngine);
     }
     
-    private UpdateResponseHeader executeMultiStatements(final DriverExecutionPrepareEngine<JDBCExecutionUnit, Connection> prepareEngine) throws SQLException {
+    private ResponseHeader executeMultiStatements(final DriverExecutionPrepareEngine<JDBCExecutionUnit, Connection> prepareEngine) throws SQLException {
         Collection<ExecutionContext> executionContexts = createExecutionContexts();
-        Map<String, List<ExecutionUnit>> dataSourcesToExecutionUnits = buildDataSourcesToExecutionUnits(executionContexts);
         ExecutionGroupContext<JDBCExecutionUnit> executionGroupContext =
-                prepareEngine.prepare(connectionSession.getUsedDatabaseName(), executionContexts.iterator().next().getRouteContext(), samplingExecutionUnit(dataSourcesToExecutionUnits),
+                prepareEngine.prepare(connectionSession.getUsedDatabaseName(), executionContexts.iterator().next().getRouteContext(), createExecutionUnits(),
                         new ExecutionGroupReportContext(connectionSession.getProcessId(), connectionSession.getUsedDatabaseName(), connectionSession.getConnectionContext().getGrantee()));
-        for (ExecutionGroup<JDBCExecutionUnit> each : executionGroupContext.getInputGroups()) {
-            for (JDBCExecutionUnit unit : each.getInputs()) {
-                prepareBatchedStatement(unit, dataSourcesToExecutionUnits);
-            }
+        batchExecutor.init(executionGroupContext);
+        executeAddBatch(executionGroupContext);
+        return new MultiStatementsUpdateResponseHeader(buildUpdateResponseHeaders(batchExecutor.executeBatch(multiSQLQueryContexts.iterator().next().getSqlStatementContext())));
+    }
+    
+    private List<ExecutionUnit> createExecutionUnits() {
+        List<ExecutionUnit> result = new ArrayList<>(batchExecutor.getBatchExecutionUnits().size());
+        for (BatchExecutionUnit each : batchExecutor.getBatchExecutionUnits()) {
+            ExecutionUnit executionUnit = each.getExecutionUnit();
+            result.add(executionUnit);
         }
-        return executeBatchedStatements(executionGroupContext);
+        return result;
     }
     
     private Collection<ExecutionContext> createExecutionContexts() {
         Collection<ExecutionContext> result = new LinkedList<>();
         for (QueryContext each : multiSQLQueryContexts) {
-            result.add(createExecutionContext(each));
-        }
-        return result;
-    }
-    
-    private Map<String, List<ExecutionUnit>> buildDataSourcesToExecutionUnits(final Collection<ExecutionContext> executionContexts) {
-        Map<String, List<ExecutionUnit>> result = new HashMap<>();
-        for (ExecutionContext each : executionContexts) {
-            for (ExecutionUnit executionUnit : each.getExecutionUnits()) {
-                result.computeIfAbsent(executionUnit.getDataSourceName(), unused -> new LinkedList<>()).add(executionUnit);
-            }
+            ExecutionContext executionContext = createExecutionContext(each);
+            batchExecutor.addBatchForExecutionUnits(executionContext.getExecutionUnits());
+            result.add(executionContext);
         }
         return result;
     }
@@ -170,58 +172,22 @@ public final class MySQLMultiStatementsHandler implements ProxyBackendHandler {
         RuleMetaData globalRuleMetaData = metaDataContexts.getMetaData().getGlobalRuleMetaData();
         ShardingSphereDatabase currentDatabase = metaDataContexts.getMetaData().getDatabase(connectionSession.getUsedDatabaseName());
         SQLAuditEngine.audit(queryContext, globalRuleMetaData, currentDatabase);
-        return kernelProcessor.generateExecutionContext(queryContext, globalRuleMetaData, metaDataContexts.getMetaData().getProps());
+        return new KernelProcessor().generateExecutionContext(queryContext, globalRuleMetaData, metaDataContexts.getMetaData().getProps());
     }
     
-    private Collection<ExecutionUnit> samplingExecutionUnit(final Map<String, List<ExecutionUnit>> dataSourcesToExecutionUnits) {
-        Collection<ExecutionUnit> result = new LinkedList<>();
-        for (List<ExecutionUnit> each : dataSourcesToExecutionUnits.values()) {
-            result.add(each.get(0));
+    private void executeAddBatch(final ExecutionGroupContext<JDBCExecutionUnit> executionGroupContext) throws SQLException {
+        for (ExecutionGroup<JDBCExecutionUnit> each : executionGroupContext.getInputGroups()) {
+            for (JDBCExecutionUnit unit : each.getInputs()) {
+                unit.getStorageResource().addBatch(unit.getExecutionUnit().getSqlUnit().getSql());
+            }
+        }
+    }
+    
+    private Collection<UpdateResponseHeader> buildUpdateResponseHeaders(final int[] updateCounts) {
+        Collection<UpdateResponseHeader> result = new LinkedList<>();
+        for (int each : updateCounts) {
+            result.add(new UpdateResponseHeader(sqlStatementSample, Collections.singletonList(new UpdateResult(each, 0L))));
         }
         return result;
-    }
-    
-    private void prepareBatchedStatement(final JDBCExecutionUnit executionUnit, final Map<String, List<ExecutionUnit>> dataSourcesToExecutionUnits) throws SQLException {
-        Statement statement = executionUnit.getStorageResource();
-        for (ExecutionUnit each : dataSourcesToExecutionUnits.get(executionUnit.getExecutionUnit().getDataSourceName())) {
-            statement.addBatch(each.getSqlUnit().getSql());
-        }
-    }
-    
-    private UpdateResponseHeader executeBatchedStatements(final ExecutionGroupContext<JDBCExecutionUnit> executionGroupContext) throws SQLException {
-        boolean isExceptionThrown = SQLExecutorExceptionHandler.isExceptionThrown();
-        ResourceMetaData resourceMetaData = metaDataContexts.getMetaData().getDatabase(connectionSession.getUsedDatabaseName()).getResourceMetaData();
-        JDBCExecutorCallback<int[]> callback = new BatchedJDBCExecutorCallback(resourceMetaData, sqlStatementSample, isExceptionThrown);
-        List<int[]> executeResults = jdbcExecutor.execute(executionGroupContext, callback);
-        int updated = 0;
-        for (int[] eachResult : executeResults) {
-            for (int each : eachResult) {
-                updated += each;
-            }
-        }
-        // TODO Each logic SQL should correspond to an OK Packet.
-        return new UpdateResponseHeader(sqlStatementSample, Collections.singletonList(new UpdateResult(updated, 0L)));
-    }
-    
-    private static final class BatchedJDBCExecutorCallback extends JDBCExecutorCallback<int[]> {
-        
-        private BatchedJDBCExecutorCallback(final ResourceMetaData resourceMetaData, final SQLStatement sqlStatement, final boolean isExceptionThrown) {
-            super(TypedSPILoader.getService(DatabaseType.class, "MySQL"), resourceMetaData, sqlStatement, isExceptionThrown);
-        }
-        
-        @Override
-        protected int[] executeSQL(final String sql, final Statement statement, final ConnectionMode connectionMode, final DatabaseType storageType) throws SQLException {
-            try {
-                return statement.executeBatch();
-            } finally {
-                statement.close();
-            }
-        }
-        
-        @SuppressWarnings("OptionalContainsCollection")
-        @Override
-        protected Optional<int[]> getSaneResult(final SQLStatement sqlStatement, final SQLException ex) {
-            return Optional.empty();
-        }
     }
 }